models/theater.py
from sqlalchemy import Column, Integer, String, Date, ForeignKey, DateTime
from sqlalchemy.orm import relationship
from sqlalchemy.ext.declarative import declarative_base

Base = declarative_base()

class Theater(Base):
    __tablename__ = 'theaters'

    id = Column(Integer, primary_key=True)
    name = Column(String)
    location = Column(String)
    capacity = Column(Integer)

    def serialize(self):
        return {
            'id': self.id,
            'name': self.name,
            'location': self.location,
            'capacity': self.capacity
        }

    def __init__(self, name, location, capacity):
        self.name = name
        self.location = location
        self.capacity = capacity


